+++ changelog.md
@@ -5,5 +5,14 @@
 ### Features Added\Updated
 * Initial Release  
 
+# dsPIC33a Bootloader and Firmware Upgrade Demo v1.0.0*
+### Release Highlights
+* Initial Release - Refer to the README.md for a summary and execution procedure
+
+### Features Added\Updated
+* Initial Release  
+
+\* **NOTE**: THIS DEMO IS AN ENGINEERING RELEASE TO DEMONSTRATE DSPIC33A BOOTLOADER FUNCTIONALITY AND SHOULD BE USED FOR REFERENCE ONLY. THIS CODE IS NOT INTENDED FOR USE IN PRODUCTION.
+